requirements.txt
transformers
faiss-cpu
nltk
torch
numpy
